debug: seperate breakpoints.test.ts into its own file
